% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{mapReduce_map}
\alias{mapReduce_map}
\title{Map Reduce Map Function}
\usage{
mapReduce_map(srcDoc, mapFunction)
}
\arguments{
\item{srcDoc}{the source document to parse}

\item{mapfunction}{the function to use, should emit a dataframe, empty if no items}
}
\value{
a list of mapped items
}
\description{
parse a map function over source data
}
\examples{
emitFunction <- function(row) {
  if(is.null(row[['Molecule']]))
    return (data.frame())
  retVal <- data.frame(molecule=row[['Molecule']],count=1)
  return (retVal)
}

list_data <- mapReduce_map(ndjson_data,emitFunction)

}
