% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{mapReduce_map_ndjson}
\alias{mapReduce_map_ndjson}
\title{Map Reduce Map Function for ndjson files}
\usage{
mapReduce_map_ndjson(srcDoc, mapFunction)
}
\arguments{
\item{srcDoc}{the source file connection to parse}

\item{mapfunction}{the function to use, should emit a dataframe, empty if no items}
}
\value{
a list of mapped items
}
\description{
parse a map function over source data
}
\examples{
file<-"data/flight_data6.jdat"
processRow <- function(row){
  if(startsWith(row,"{")){
    json_final<-fromJSON(row)
    if(!is.null(json_final[["flightData"]])){
      timeStamp<-as.numeric(as.POSIXct(strptime(json_final[["time"]], "\%Y/\%m/\%d \%H:\%M:\%S")))
      if(timeStamp>=initial_time)
        return(data.frame(json_final[["flightData"]],
               time=json_final[["time"]],timestamp=((timeStamp-initial_time)/60)))
      else
        return(data.frame())
    }
    else
      return(data.frame())
  } else {
    return(data.frame())
  }
}
con = file(file, "r")
list_data <- mapReduce_map_ndjson(con,processRow) 
close(con)

}
