% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{mapReduce_numWorkers}
\alias{mapReduce_numWorkers}
\title{Map Reduce update number of worker}
\usage{
mapReduce_numWorkers(numWorkers)
}
\arguments{
\item{numWorkers}{the number of workers}
}
\value{
void
}
\description{
By default, detectCores is used to set the number of workers, use this function to
modify this value.
}
