\name{mapReduce-package}
\alias{mapReduce-package}
\docType{package}
\title{ mapReduce - flexible mapReduce algorithm for parallel computation }
\description{                       

  mapReduce is an algorithm provides a simple framework for parallel
  computations.  This implementation provides the following features:
    * pure R
    * simple R-style syntax
    * agnostic to parallelization backend
    
  mapReduce is also a convenient replacement for \code{\link{by}} 
  and \code{\link{aggregate}}.  

}
\details{
\tabular{ll}{
Package: \tab mapReduce\cr
Type: \tab Package\cr
Version: \tab 1.2.5\cr
Date: \tab 2012-04-29\cr
License: \tab GPL (>=2 )\cr
LazyLoad: \tab yes\cr
}

All examples are in \code{\link{mapReduce}}

}
\author{
Christopher Brown

Maintainer: Christopher Brown <chris.brown@decisionpatterns.com>
}

\references{
Dean and Gemawatt, (2004) MapReduce: Simplified Data Processing on Large Clusters. OSDI'04: Sixth Symposium on Operating System Design and Implementation.  

also: \url{http://labs.google.com/papers/mapreduce.html}

}

\keyword{ package }

